Move session tests to their own module
